+++ packages/apollo-datasource-rest/src/RESTDataSource.ts
@@ -104,7 +104,11 @@ export abstract class RESTDataSource<TContext = any> extends DataSource {
 
   protected parseBody(response: Response): Promise<object | string> {
     const contentType = response.headers.get('Content-Type');
-    if (contentType && contentType.startsWith('application/json')) {
+    if (
+      contentType &&
+      (contentType.startsWith('application/json') ||
+        contentType.startsWith('application/hal+json'))
+    ) {
       return response.json();
     } else {
       return response.text();