+++ packages/apollo-server-core/src/ApolloServer.ts
@@ -14,6 +14,7 @@ import { GraphQLExtension } from 'graphql-extensions';
 import { EngineReportingAgent } from 'apollo-engine-reporting';
 import { InMemoryLRUCache } from 'apollo-server-caching';
 import { ApolloServerPlugin } from 'apollo-server-plugin-base';
+import runtimeSupportsUploads from './utils/runtimeSupportsUploads';
 
 import {
   SubscriptionServer,
@@ -43,6 +44,16 @@ import {
   createPlaygroundOptions,
   PlaygroundRenderPageOptions,
 } from './playground';
+
+import { generateSchemaHash } from './utils/schemaHash';
+import {
+  processGraphQLRequest,
+  GraphQLRequestContext,
+  GraphQLRequest,
+} from './requestPipeline';
+
+import { Headers } from 'apollo-server-env';
+import { buildServiceDefinition } from '@apollographql/apollo-tools';
 
 const NoIntrospection = (context: ValidationContext) => ({
   Field(node: FieldDefinitionNode) {
@@ -78,6 +89,9 @@ function getEngineServiceId(engine: Config['engine']): string | undefined {
   return;
 }
 
+const forbidUploadsForTesting =
+  process && process.env.NODE_ENV === 'test' && !runtimeSupportsUploads;
+
 export class ApolloServerBase {
   public subscriptionsPath?: string;
   public graphqlPath: string = '/graphql';
@@ -87,6 +101,7 @@ export class ApolloServerBase {
   private engineReportingAgent?: EngineReportingAgent;
   private engineServiceId?: string;
   private extensions: Array<() => GraphQLExtension>;
+  private schemaHash: string;
   protected plugins: ApolloServerPlugin[] = [];
 
   protected schema: GraphQLSchema;
@@ -107,6 +122,7 @@ export class ApolloServerBase {
       resolvers,
       schema,
       schemaDirectives,
+      modules,
       typeDefs,
       introspection,
       mocks,
@@ -187,8 +203,16 @@ export class ApolloServerBase {
     this.requestOptions = requestOptions as GraphQLOptions;
     this.context = context;
 
-    if (uploads !== false) {
+    if (uploads !== false && !forbidUploadsForTesting) {
       if (this.supportsUploads()) {
+        if (!runtimeSupportsUploads) {
+          printNodeFileUploadsMessage();
+          throw new Error(
+            '`graphql-upload` is no longer supported on Node.js < v8.5.0.  ' +
+              'See https://bit.ly/gql-upload-node-6.',
+          );
+        }
+
         if (uploads === true || typeof uploads === 'undefined') {
           this.uploadsConfig = {};
         } else {
@@ -198,17 +222,23 @@ export class ApolloServerBase {
         //default we enable them if supported by the integration
       } else if (uploads) {
         throw new Error(
-          'This implementation of ApolloServer does not support file uploads because the environmnet cannot accept multi-part forms',
+          'This implementation of ApolloServer does not support file uploads because the environment cannot accept multi-part forms',
         );
       }
     }
 
     if (schema) {
       this.schema = schema;
+    } else if (modules) {
+      const { schema, errors } = buildServiceDefinition(modules);
+      if (errors && errors.length > 0) {
+        throw new Error(errors.map(error => error.message).join('\n\n'));
+      }
+      this.schema = schema!;
     } else {
       if (!typeDefs) {
         throw Error(
-          'Apollo Server requires either an existing schema or typeDefs',
+          'Apollo Server requires either an existing schema, modules or typeDefs',
         );
       }
 
@@ -231,9 +261,7 @@ export class ApolloServerBase {
       );
 
       if (this.uploadsConfig) {
-        const {
-          GraphQLUpload,
-        } = require('@apollographql/apollo-upload-server');
+        const { GraphQLUpload } = require('graphql-upload');
         if (resolvers && !resolvers.Upload) {
           resolvers.Upload = GraphQLUpload;
         }
@@ -254,7 +282,7 @@ export class ApolloServerBase {
       });
     }
 
-    if (mocks || typeof mockEntireSchema !== 'undefined') {
+    if (mocks || (typeof mockEntireSchema !== 'undefined' && mocks !== false)) {
       addMockFunctionsToSchema({
         schema: this.schema,
         mocks:
@@ -265,6 +293,10 @@ export class ApolloServerBase {
           typeof mockEntireSchema === 'undefined' ? false : !mockEntireSchema,
       });
     }
+
+    // The schema hash is a string representation of the shape of the schema
+    // it is used for reporting and can be used for a cache key if needed
+    this.schemaHash = generateSchemaHash(this.schema);
 
     // Note: doRunQuery will add its own extensions if you set tracing,
     // or cacheControl.
@@ -290,6 +322,13 @@ export class ApolloServerBase {
     if (this.engineServiceId) {
       this.engineReportingAgent = new EngineReportingAgent(
         typeof engine === 'object' ? engine : Object.create(null),
+        {
+          schema: this.schema,
+          schemaHash: this.schemaHash,
+          engine: {
+            serviceID: this.engineServiceId,
+          },
+        },
       );
       // Let's keep this extension second so it wraps everything, except error formatting
       this.extensions.push(() => this.engineReportingAgent!.newExtension());
@@ -341,6 +380,7 @@ export class ApolloServerBase {
           plugin.serverWillStart &&
           plugin.serverWillStart({
             schema: this.schema,
+            schemaHash: this.schemaHash,
             engine: {
               serviceID: this.engineServiceId,
             },
@@ -438,15 +478,11 @@ export class ApolloServerBase {
       return;
     }
 
-    // FIXME: We also want to support default exports and possibly module names
-    // but this requires adjustments to typing (see PluginDefinition type), and
-    // I had to give up on that for now.
     this.plugins = plugins.map(plugin => {
       if (typeof plugin === 'function') {
-        return new plugin();
-      } else {
-        return plugin as ApolloServerPlugin;
+        return plugin();
       }
+      return plugin;
     });
   }
 
@@ -487,4 +523,61 @@ export class ApolloServerBase {
       ...this.requestOptions,
     } as GraphQLOptions;
   }
+
+  public async executeOperation(request: GraphQLRequest) {
+    let options;
+
+    try {
+      options = await this.graphQLServerOptions();
+    } catch (e) {
+      e.message = `Invalid options provided to ApolloServer: ${e.message}`;
+      throw new Error(e);
+    }
+
+    if (typeof options.context === 'function') {
+      options.context = (options.context as () => never)();
+    }
+
+    const requestCtx: GraphQLRequestContext = {
+      request,
+      context: options.context || Object.create(null),
+      cache: options.cache!,
+      response: {
+        http: {
+          headers: new Headers(),
+        },
+      },
+    };
+
+    return processGraphQLRequest(options, requestCtx);
+  }
+}
+
+function printNodeFileUploadsMessage() {
+  console.error(
+    [
+      '*****************************************************************',
+      '*                                                               *',
+      '* ERROR! Manual intervention is necessary for Node.js < v8.5.0! *',
+      '*                                                               *',
+      '*****************************************************************',
+      '',
+      'The third-party `graphql-upload` package, which is used to implement',
+      'file uploads in Apollo Server 2.x, no longer supports Node.js LTS',
+      'versions prior to Node.js v8.5.0.',
+      '',
+      'Deployments which NEED file upload capabilities should update to',
+      'Node.js >= v8.5.0 to continue using uploads.',
+      '',
+      'If this server DOES NOT NEED file uploads and wishes to continue',
+      'using this version of Node.js, uploads can be disabled by adding:',
+      '',
+      '  uploads: false,',
+      '',
+      '...to the options for Apollo Server and re-deploying the server.',
+      '',
+      'For more information, see https://bit.ly/gql-upload-node-6.',
+      '',
+    ].join('\n'),
+  );
 }