+++ packages/apollo-server-express/src/__tests__/ApolloServer.test.ts
@@ -11,7 +11,7 @@ import { gql, AuthenticationError, Config } from 'apollo-server-core';
 import { ApolloServer, ServerRegistration } from '../ApolloServer';
 
 import {
-  atLeastMajorNodeVersion,
+  NODE_MAJOR_VERSION,
   testApolloServer,
   createServerInfo,
 } from 'apollo-server-integration-testsuite';
@@ -427,8 +427,9 @@ describe('apollo-server-express', () => {
         });
       });
     });
-    // NODE: Intentionally skip file upload tests on Node.js 10 or higher.
-    (atLeastMajorNodeVersion(10) ? describe.skip : describe)(
+    // NODE: Skip Node.js 6, but only because `graphql-upload`
+    // doesn't support it.
+    (NODE_MAJOR_VERSION === 6 ? describe.skip : describe)(
       'file uploads',
       () => {
         it('enabled uploads', async () => {